[*] fixed double variable inconsistent value

+++ app/src/main/java/com/sleepybear/mymoviecatalogue/models/nowplaying/NowplayingResult.java
@@ -17,7 +17,7 @@ public class NowplayingResult {
     private Boolean video;
     @SerializedName("vote_average")
     @Expose
-    private Integer voteAverage;
+    private Double voteAverage;
     @SerializedName("title")
     @Expose
     private String title;
@@ -73,11 +73,11 @@ public class NowplayingResult {
         this.video = video;
     }
 
-    public Integer getVoteAverage() {
+    public Double getVoteAverage() {
         return voteAverage;
     }
 
-    public void setVoteAverage(Integer voteAverage) {
+    public void setVoteAverage(Double voteAverage) {
         this.voteAverage = voteAverage;
     }
 
@@ -160,5 +160,4 @@ public class NowplayingResult {
     public void setReleaseDate(String releaseDate) {
         this.releaseDate = releaseDate;
     }
-
 }
